# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Operator - personal life OS
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DailyCheckin {
  id: number;
  date: string;
  notes: string;
  energyLevel: number;
  focusLevel: number;
  healthLevel: number;
  sleepQuality: number;
  mood: number;
  tasksCompleted: string;
  tasksMissed: string;
  /** @nullable */
  habitsCompleted: string | null;
  /** @nullable */
  symptomsNotes: string | null;
  /** @nullable */
  claudeReflection: string | null;
  /** @nullable */
  openaiPlan: string | null;
  /** @nullable */
  combinedAdvice: string | null;
  createdAt: string;
}

export interface CreateDailyCheckinBody {
  date: string;
  notes: string;
  energyLevel: number;
  focusLevel: number;
  healthLevel: number;
  sleepQuality: number;
  mood: number;
  tasksCompleted: string;
  tasksMissed: string;
  /** @nullable */
  habitsCompleted?: string | null;
  /** @nullable */
  symptomsNotes?: string | null;
}

export interface WeeklyReview {
  id: number;
  weekStartDate: string;
  weekNotes: string;
  mainWins: string;
  mainFrustrations: string;
  energyTrend: string;
  healthTrend: string;
  goalsNextWeek: string;
  existingCommitments: string;
  /** @nullable */
  claudeReflection: string | null;
  /** @nullable */
  openaiPlan: string | null;
  /** @nullable */
  combinedAdvice: string | null;
  createdAt: string;
}

export interface CreateWeeklyReviewBody {
  weekStartDate: string;
  weekNotes: string;
  mainWins: string;
  mainFrustrations: string;
  energyTrend: string;
  healthTrend: string;
  goalsNextWeek: string;
  existingCommitments: string;
}

export interface OperatorStats {
  totalCheckins: number;
  totalWeeklyReviews: number;
  /** @nullable */
  avgEnergy: number | null;
  /** @nullable */
  avgFocus: number | null;
  /** @nullable */
  avgMood: number | null;
  /** @nullable */
  avgSleep: number | null;
  /** @nullable */
  avgHealth: number | null;
  /** @nullable */
  lastCheckinDate: string | null;
  /** @nullable */
  lastWeeklyReviewDate: string | null;
  streakDays: number;
}

export interface TrendPoint {
  date: string;
  energy: number;
  focus: number;
  mood: number;
  sleep: number;
  health: number;
}

export type ListDailyCheckinsParams = {
  limit?: number;
  offset?: number;
};

export type ListWeeklyReviewsParams = {
  limit?: number;
  offset?: number;
};