# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Operator - personal life OS
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all daily check-ins
 */
export const listDailyCheckinsQueryLimitDefault = 50;
export const listDailyCheckinsQueryOffsetDefault = 0;

export const ListDailyCheckinsQueryParams = zod.object({
  limit: zod.coerce.number().default(listDailyCheckinsQueryLimitDefault),
  offset: zod.coerce.number().default(listDailyCheckinsQueryOffsetDefault),
});

export const ListDailyCheckinsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  notes: zod.string(),
  energyLevel: zod.number(),
  focusLevel: zod.number(),
  healthLevel: zod.number(),
  sleepQuality: zod.number(),
  mood: zod.number(),
  tasksCompleted: zod.string(),
  tasksMissed: zod.string(),
  habitsCompleted: zod.string().nullable(),
  symptomsNotes: zod.string().nullable(),
  claudeReflection: zod.string().nullable(),
  openaiPlan: zod.string().nullable(),
  combinedAdvice: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListDailyCheckinsResponse = zod.array(
  ListDailyCheckinsResponseItem,
);

/**
 * @summary Create a daily check-in with AI analysis
 */
export const CreateDailyCheckinBody = zod.object({
  date: zod.string(),
  notes: zod.string(),
  energyLevel: zod.number(),
  focusLevel: zod.number(),
  healthLevel: zod.number(),
  sleepQuality: zod.number(),
  mood: zod.number(),
  tasksCompleted: zod.string(),
  tasksMissed: zod.string(),
  habitsCompleted: zod.string().nullish(),
  symptomsNotes: zod.string().nullish(),
});

/**
 * @summary Get a specific daily check-in
 */
export const GetDailyCheckinParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDailyCheckinResponse = zod.object({
  id: zod.number(),
  date: zod.string(),
  notes: zod.string(),
  energyLevel: zod.number(),
  focusLevel: zod.number(),
  healthLevel: zod.number(),
  sleepQuality: zod.number(),
  mood: zod.number(),
  tasksCompleted: zod.string(),
  tasksMissed: zod.string(),
  habitsCompleted: zod.string().nullable(),
  symptomsNotes: zod.string().nullable(),
  claudeReflection: zod.string().nullable(),
  openaiPlan: zod.string().nullable(),
  combinedAdvice: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a daily check-in
 */
export const DeleteDailyCheckinParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all weekly reviews
 */
export const listWeeklyReviewsQueryLimitDefault = 20;
export const listWeeklyReviewsQueryOffsetDefault = 0;

export const ListWeeklyReviewsQueryParams = zod.object({
  limit: zod.coerce.number().default(listWeeklyReviewsQueryLimitDefault),
  offset: zod.coerce.number().default(listWeeklyReviewsQueryOffsetDefault),
});

export const ListWeeklyReviewsResponseItem = zod.object({
  id: zod.number(),
  weekStartDate: zod.string(),
  weekNotes: zod.string(),
  mainWins: zod.string(),
  mainFrustrations: zod.string(),
  energyTrend: zod.string(),
  healthTrend: zod.string(),
  goalsNextWeek: zod.string(),
  existingCommitments: zod.string(),
  claudeReflection: zod.string().nullable(),
  openaiPlan: zod.string().nullable(),
  combinedAdvice: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListWeeklyReviewsResponse = zod.array(
  ListWeeklyReviewsResponseItem,
);

/**
 * @summary Create a weekly review with AI analysis
 */
export const CreateWeeklyReviewBody = zod.object({
  weekStartDate: zod.string(),
  weekNotes: zod.string(),
  mainWins: zod.string(),
  mainFrustrations: zod.string(),
  energyTrend: zod.string(),
  healthTrend: zod.string(),
  goalsNextWeek: zod.string(),
  existingCommitments: zod.string(),
});

/**
 * @summary Get a specific weekly review
 */
export const GetWeeklyReviewParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWeeklyReviewResponse = zod.object({
  id: zod.number(),
  weekStartDate: zod.string(),
  weekNotes: zod.string(),
  mainWins: zod.string(),
  mainFrustrations: zod.string(),
  energyTrend: zod.string(),
  healthTrend: zod.string(),
  goalsNextWeek: zod.string(),
  existingCommitments: zod.string(),
  claudeReflection: zod.string().nullable(),
  openaiPlan: zod.string().nullable(),
  combinedAdvice: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a weekly review
 */
export const DeleteWeeklyReviewParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get overall stats summary
 */
export const GetOperatorStatsResponse = zod.object({
  totalCheckins: zod.number(),
  totalWeeklyReviews: zod.number(),
  avgEnergy: zod.number().nullable(),
  avgFocus: zod.number().nullable(),
  avgMood: zod.number().nullable(),
  avgSleep: zod.number().nullable(),
  avgHealth: zod.number().nullable(),
  lastCheckinDate: zod.string().nullable(),
  lastWeeklyReviewDate: zod.string().nullable(),
  streakDays: zod.number(),
});

/**
 * @summary Get recent metric trends (last 7 check-ins)
 */
export const GetOperatorTrendsResponseItem = zod.object({
  date: zod.string(),
  energy: zod.number(),
  focus: zod.number(),
  mood: zod.number(),
  sleep: zod.number(),
  health: zod.number(),
});
export const GetOperatorTrendsResponse = zod.array(
  GetOperatorTrendsResponseItem,
);