# lib/api-zod/src/generated/types/createDailyCheckinBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Operator - personal life OS
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDailyCheckinBody {
  date: string;
  notes: string;
  energyLevel: number;
  focusLevel: number;
  healthLevel: number;
  sleepQuality: number;
  mood: number;
  tasksCompleted: string;
  tasksMissed: string;
  /** @nullable */
  habitsCompleted?: string | null;
  /** @nullable */
  symptomsNotes?: string | null;
}