# lib/api-zod/src/generated/types/createWeeklyReviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Operator - personal life OS
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWeeklyReviewBody {
  weekStartDate: string;
  weekNotes: string;
  mainWins: string;
  mainFrustrations: string;
  energyTrend: string;
  healthTrend: string;
  goalsNextWeek: string;
  existingCommitments: string;
}